First Wwise commit
Hoping that GitLFS will allow for the soundbanks to upload. It would be great if this works as it is way faster to implement clean sounds. Just have a really crappy player movement speed in at the moment along with some BG sound. Still need to code the rotate and pick up sounds